refactor: Extract testable logic to utils and import in tests
Move parseToolUsage to utils/terminalUtils.ts and height constants
to utils/inputBarConstants.ts so tests import from source instead
of duplicating logic.
