+++ apps/mobile/src/utils/inputBarConstants.ts
@@ -1,0 +1,2 @@
+export const MIN_INPUT_HEIGHT = 36;
+export const MAX_INPUT_HEIGHT = 120;
